Initial Implementation of Module Enforcer
Creates an initial version of the project that allows you
to use the @Module and @RequiredMethod annotations in Java
classes to make it easier to split up your activity logic
and have static time analysis that required logic is called.

The Readme is just a draft and the library doesn't work for
Kotlin (yet).